# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Loan Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: loans
    description: Loan management
  - name: payments
    description: Payment tracking
  - name: dashboard
    description: Dashboard summaries
  - name: profile
    description: Global financial profile
  - name: settings
    description: User account preferences
  - name: feedback
    description: User feedback and feature requests

paths:
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the current user's global financial profile
      responses:
        "200":
          description: Financial profile (defaults when none saved yet)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Create or replace the current user's global financial profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialProfileData"
      responses:
        "200":
          description: Saved financial profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialProfile"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get the current user's account preferences
      responses:
        "200":
          description: User settings (defaults when none saved yet)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "401":
          description: Unauthorized
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Create or replace the current user's account preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettingsData"
      responses:
        "200":
          description: Saved settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized

  /feedback:
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit user feedback, a rating, or a feature request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized

  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List all loans for current user
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, paid, overdue]
          required: false
      responses:
        "200":
          description: List of loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
        "401":
          description: Unauthorized
    post:
      operationId: createLoan
      tags: [loans]
      summary: Create a new loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Loan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized

  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get a loan by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Not found
        "401":
          description: Unauthorized
    patch:
      operationId: updateLoan
      tags: [loans]
      summary: Update a loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanUpdate"
      responses:
        "200":
          description: Updated loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Not found
        "401":
          description: Unauthorized
    delete:
      operationId: deleteLoan
      tags: [loans]
      summary: Delete a loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
        "401":
          description: Unauthorized

  /loans/{id}/payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments for a loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
        "401":
          description: Unauthorized
    post:
      operationId: addPayment
      tags: [payments]
      summary: Add a payment to a loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized

  /loans/{id}/payments/{paymentId}:
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: paymentId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
        "401":
          description: Unauthorized

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Unauthorized

  /loans/recent:
    get:
      operationId: getRecentLoans
      tags: [loans]
      summary: Get recent loans (last 5)
      responses:
        "200":
          description: Recent loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RateChange:
      type: object
      required:
        - effectiveDate
        - newRate
      properties:
        effectiveDate:
          type: string
        newRate:
          type: number
          minimum: 0

    Loan:
      type: object
      required:
        - id
        - userId
        - borrowerName
        - principalAmount
        - interestRate
        - startDate
        - dueDate
        - status
        - totalPaid
        - remainingAmount
        - createdAt
        - rateChanges
      properties:
        id:
          type: integer
        userId:
          type: string
        borrowerName:
          type: string
        principalAmount:
          type: number
        interestRate:
          type: number
        tenureMonths:
          type: ["integer", "null"]
        startDate:
          type: string
        dueDate:
          type: string
        bank:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paid, overdue]
        totalPaid:
          type: number
        remainingAmount:
          type: number
        createdAt:
          type: string
        rateChanges:
          type: array
          items:
            $ref: "#/components/schemas/RateChange"

    LoanInput:
      type: object
      required:
        - borrowerName
        - principalAmount
        - interestRate
      properties:
        borrowerName:
          type: string
          minLength: 1
        principalAmount:
          type: number
          minimum: 0
        interestRate:
          type: number
          minimum: 0
        tenureMonths:
          type: integer
          minimum: 1
        startDate:
          type: string
        dueDate:
          type: string
        bank:
          type: string
        description:
          type: string
        rateChanges:
          type: array
          items:
            $ref: "#/components/schemas/RateChange"

    LoanUpdate:
      type: object
      properties:
        borrowerName:
          type: string
        principalAmount:
          type: number
        interestRate:
          type: number
        tenureMonths:
          type: integer
          minimum: 1
        startDate:
          type: string
        dueDate:
          type: string
        bank:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, paid, overdue]
        rateChanges:
          type: array
          items:
            $ref: "#/components/schemas/RateChange"

    Payment:
      type: object
      required:
        - id
        - loanId
        - amount
        - paymentDate
        - createdAt
      properties:
        id:
          type: integer
        loanId:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    PaymentInput:
      type: object
      required:
        - amount
        - paymentDate
      properties:
        amount:
          type: number
          minimum: 0
        paymentDate:
          type: string
        notes:
          type: string

    ProfileDebtItem:
      type: object
      required:
        - id
        - name
        - balance
        - rate
        - minPayment
      properties:
        id:
          type: string
        name:
          type: string
        balance:
          type: number
        rate:
          type: number
        minPayment:
          type: number

    FinancialProfileData:
      type: object
      required:
        - name
        - age
        - occupation
        - monthlyIncome
        - additionalIncome
        - rent
        - emi
        - insurance
        - utilities
        - schoolFees
        - internet
        - otherFixed
        - food
        - fuel
        - travel
        - entertainment
        - shopping
        - medical
        - miscellaneous
        - currentSavings
        - existingInvestments
        - creditCardDebt
        - debts
        - goals
        - riskProfile
      properties:
        name:
          type: string
        age:
          type: number
        occupation:
          type: string
        monthlyIncome:
          type: number
        additionalIncome:
          type: number
        rent:
          type: number
        emi:
          type: number
        insurance:
          type: number
        utilities:
          type: number
        schoolFees:
          type: number
        internet:
          type: number
        otherFixed:
          type: number
        food:
          type: number
        fuel:
          type: number
        travel:
          type: number
        entertainment:
          type: number
        shopping:
          type: number
        medical:
          type: number
        miscellaneous:
          type: number
        currentSavings:
          type: number
        existingInvestments:
          type: number
        creditCardDebt:
          type: number
        debts:
          type: array
          items:
            $ref: "#/components/schemas/ProfileDebtItem"
        goals:
          type: array
          items:
            type: string
        riskProfile:
          type: string
          enum: [conservative, moderate, aggressive]

    FinancialProfile:
      type: object
      required:
        - data
        - updatedAt
      properties:
        data:
          $ref: "#/components/schemas/FinancialProfileData"
        updatedAt:
          type: ["string", "null"]

    UserNotificationPrefs:
      type: object
      required:
        - emiReminder
        - dueDateReminder
        - prepaymentReminder
        - weeklySummary
        - monthlySummary
        - emailNotifications
        - pushNotifications
      properties:
        emiReminder:
          type: boolean
        dueDateReminder:
          type: boolean
        prepaymentReminder:
          type: boolean
        weeklySummary:
          type: boolean
        monthlySummary:
          type: boolean
        emailNotifications:
          type: boolean
        pushNotifications:
          type: boolean

    UserSettingsData:
      type: object
      required:
        - currency
        - locale
        - dateFormat
        - defaultInterestRate
        - defaultTenureMonths
        - autoSaveCalculations
        - notifications
      properties:
        currency:
          type: string
          description: ISO 4217 currency code, e.g. INR, USD
        locale:
          type: string
          description: BCP-47 locale for number formatting, e.g. en-IN
        dateFormat:
          type: string
          enum: [DD/MM/YYYY, MM/DD/YYYY, YYYY-MM-DD]
        defaultInterestRate:
          type: number
        defaultTenureMonths:
          type: number
        autoSaveCalculations:
          type: boolean
        notifications:
          $ref: "#/components/schemas/UserNotificationPrefs"

    UserSettings:
      type: object
      required:
        - data
        - updatedAt
      properties:
        data:
          $ref: "#/components/schemas/UserSettingsData"
        updatedAt:
          type: ["string", "null"]

    FeedbackInput:
      type: object
      required:
        - kind
        - message
      properties:
        kind:
          type: string
          enum: [rating, feedback, feature]
        rating:
          type: number
        message:
          type: string
          minLength: 1
        email:
          type: string

    Feedback:
      type: object
      required:
        - id
        - kind
        - message
        - createdAt
      properties:
        id:
          type: integer
        kind:
          type: string
        rating:
          type: ["number", "null"]
        message:
          type: string
        email:
          type: ["string", "null"]
        createdAt:
          type: string

    DashboardSummary:
      type: object
      required:
        - totalLoans
        - activeLoans
        - overdueLoans
        - paidLoans
        - totalLent
        - totalCollected
        - totalOutstanding
      properties:
        totalLoans:
          type: integer
        activeLoans:
          type: integer
        overdueLoans:
          type: integer
        paidLoans:
          type: integer
        totalLent:
          type: number
        totalCollected:
          type: number
        totalOutstanding:
          type: number