# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loan Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Feedback {
  id: number;
  kind: string;
  /** @nullable */
  rating?: number | null;
  message: string;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}