# lib/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loan Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackInputKind } from './feedbackInputKind';

export interface FeedbackInput {
  kind: FeedbackInputKind;
  rating?: number;
  /** @minLength 1 */
  message: string;
  email?: string;
}