# lib/api-zod/src/generated/types/userNotificationPrefs.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loan Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface UserNotificationPrefs {
  emiReminder: boolean;
  dueDateReminder: boolean;
  prepaymentReminder: boolean;
  weeklySummary: boolean;
  monthlySummary: boolean;
  emailNotifications: boolean;
  pushNotifications: boolean;
}