# lib/api-zod/src/generated/types/userSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loan Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSettingsData } from './userSettingsData';

export interface UserSettings {
  data: UserSettingsData;
  /** @nullable */
  updatedAt: string | null;
}