# lib/api-zod/src/generated/types/userSettingsData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Loan Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { UserNotificationPrefs } from './userNotificationPrefs';
import type { UserSettingsDataDateFormat } from './userSettingsDataDateFormat';

export interface UserSettingsData {
  /** ISO 4217 currency code, e.g. INR, USD */
  currency: string;
  /** BCP-47 locale for number formatting, e.g. en-IN */
  locale: string;
  dateFormat: UserSettingsDataDateFormat;
  defaultInterestRate: number;
  defaultTenureMonths: number;
  autoSaveCalculations: boolean;
  notifications: UserNotificationPrefs;
}